+++ plugins/postcss-custom-selectors/src/is-processable-custom-selector-rule.ts
@@ -1,6 +1,6 @@
 import type { AtRule, ChildNode, Container, Document } from 'postcss';
 
-const allowedParentAtRules = new Set(['scope', 'container']);
+const allowedParentAtRules = new Set(['scope', 'container', 'layer']);
 
 export function isProcessableCustomSelectorRule(atRule: AtRule): boolean {
 	if (atRule.type !== 'atrule') {