+++ plugins/postcss-custom-selectors/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Custom Selectors
 
+### Unreleased
+
+- Added: Support for Cascade Layers.
+
 ### 7.0.0 (November 14, 2022)
 
 - Updated: Support for Node v14+ (major).